backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
python-dotenv==1.0.1
google-genai==1.3.0
beautifulsoup4==4.12.3
playwright==1.49.0